perf: 优化角色管理模块service层及异常处理

+++ ruoyi-fastapi-backend/module_admin/service/role_service.py
@@ -2,6 +2,8 @@ from module_admin.entity.vo.user_vo import UserInfoModel, UserRolePageQueryModel
 from module_admin.entity.vo.common_vo import CrudResponseModel
 from module_admin.dao.user_dao import UserDao
 from module_admin.dao.role_dao import *
+from config.constant import CommonConstant
+from exceptions.exception import ServiceException
 from utils.page_util import PageResponseModel
 from utils.common_util import export_list2excel, CamelCaseUtil
 
@@ -30,7 +32,8 @@ class RoleService:
         :param role_id: 角色id
         :return: 当前角色id的部门树信息对象
         """
-        role_dept_list = await RoleDao.get_role_dept_dao(query_db, role_id)
+        role = await cls.role_detail_services(query_db, role_id)
+        role_dept_list = await RoleDao.get_role_dept_dao(query_db, role)
         checked_keys = [row.dept_id for row in role_dept_list]
         result = RoleDeptQueryModel(
             checkedKeys=checked_keys
@@ -39,17 +42,73 @@ class RoleService:
         return result
 
     @classmethod
-    async def get_role_list_services(cls, query_db: AsyncSession, query_object: RolePageQueryModel, is_page: bool = False):
+    async def get_role_list_services(cls, query_db: AsyncSession, query_object: RolePageQueryModel, data_scope_sql: str, is_page: bool = False):
         """
         获取角色列表信息service
         :param query_db: orm对象
         :param query_object: 查询参数对象
+        :param data_scope_sql: 数据权限对应的查询sql语句
         :param is_page: 是否开启分页
         :return: 角色列表信息对象
         """
-        role_list_result = await RoleDao.get_role_list(query_db, query_object, is_page)
+        role_list_result = await RoleDao.get_role_list(query_db, query_object, data_scope_sql, is_page)
 
         return role_list_result
+
+    @classmethod
+    async def check_role_allowed_services(cls, check_role: RoleModel):
+        """
+        校验角色是否允许操作service
+        :param check_role: 角色信息
+        :return: 校验结果
+        """
+        if check_role.admin:
+            raise ServiceException(message='不允许操作超级管理员角色')
+        else:
+            return CrudResponseModel(is_success=True, message='校验通过')
+
+    @classmethod
+    async def check_role_data_scope_services(cls, query_db: AsyncSession, role_id: int, data_scope_sql: str):
+        """
+        校验角色是否有数据权限service
+        :param query_db: orm对象
+        :param role_id: 角色id
+        :param data_scope_sql: 数据权限对应的查询sql语句
+        :return: 校验结果
+        """
+        roles = await RoleDao.get_role_list(query_db, RolePageQueryModel(roleId=role_id), data_scope_sql, is_page=False)
+        if roles:
+            return CrudResponseModel(is_success=True, message='校验通过')
+        else:
+            raise ServiceException(message='没有权限访问角色数据')
+
+    @classmethod
+    async def check_role_name_unique_services(cls, query_db: AsyncSession, page_object: RoleModel):
+        """
+        校验角色名称是否唯一service
+        :param query_db: orm对象
+        :param page_object: 角色对象
+        :return: 校验结果
+        """
+        role_id = -1 if page_object.role_id is None else page_object.role_id
+        role = await RoleDao.get_role_by_info(query_db, RoleModel(roleName=page_object.role_name))
+        if role and role.role_id != role_id:
+            return CommonConstant.NOT_UNIQUE
+        return CommonConstant.UNIQUE
+
+    @classmethod
+    async def check_role_key_unique_services(cls, query_db: AsyncSession, page_object: RoleModel):
+        """
+        校验角色权限字符是否唯一service
+        :param query_db: orm对象
+        :param page_object: 角色对象
+        :return: 校验结果
+        """
+        role_id = -1 if page_object.role_id is None else page_object.role_id
+        role = await RoleDao.get_role_by_info(query_db, RoleModel(roleKey=page_object.role_key))
+        if role and role.role_id != role_id:
+            return CommonConstant.NOT_UNIQUE
+        return CommonConstant.UNIQUE
 
     @classmethod
     async def add_role_services(cls, query_db: AsyncSession, page_object: AddRoleModel):
@@ -60,12 +119,10 @@ class RoleService:
         :return: 新增角色校验结果
         """
         add_role = RoleModel(**page_object.model_dump(by_alias=True))
-        role_name = await RoleDao.get_role_by_info(query_db, RoleModel(roleName=page_object.role_name))
-        role_key = await RoleDao.get_role_by_info(query_db, RoleModel(roleKey=page_object.role_key))
-        if role_name:
-            result = dict(is_success=False, message='角色名称已存在')
-        elif role_key:
-            result = dict(is_success=False, message='权限字符已存在')
+        if not await cls.check_role_name_unique_services(query_db, page_object):
+            raise ServiceException(message=f'新增角色{page_object.post_name}失败，角色名称已存在')
+        elif not await cls.check_role_key_unique_services(query_db, page_object):
+            raise ServiceException(message=f'新增角色{page_object.post_name}失败，角色权限已存在')
         else:
             try:
                 add_result = await RoleDao.add_role_dao(query_db, add_role)
@@ -74,12 +131,10 @@ class RoleService:
                     for menu in page_object.menu_ids:
                         await RoleDao.add_role_menu_dao(query_db, RoleMenuModel(roleId=role_id, menuId=menu))
                 await query_db.commit()
-                result = dict(is_success=True, message='新增成功')
+                return CrudResponseModel(is_success=True, message='新增成功')
             except Exception as e:
                 await query_db.rollback()
                 raise e
-
-        return CrudResponseModel(**result)
 
     @classmethod
     async def edit_role_services(cls, query_db: AsyncSession, page_object: AddRoleModel):
@@ -96,16 +151,11 @@ class RoleService:
             del edit_role['type']
         role_info = await cls.role_detail_services(query_db, edit_role.get('role_id'))
         if role_info:
-            if page_object.type != 'status' and role_info.role_name != page_object.role_name:
-                role_name = await RoleDao.get_role_by_info(query_db, RoleModel(roleName=page_object.role_name))
-                if role_name:
-                    result = dict(is_success=False, message='角色名称已存在')
-                    return CrudResponseModel(**result)
-            elif page_object.type != 'status' and role_info.role_key != page_object.role_key:
-                role_key = await RoleDao.get_role_by_info(query_db, RoleModel(roleKey=page_object.role_key))
-                if role_key:
-                    result = dict(is_success=False, message='权限字符已存在')
-                    return CrudResponseModel(**result)
+            if page_object.type != 'status':
+                if not await cls.check_role_name_unique_services(query_db, page_object):
+                    raise ServiceException(message=f'修改角色{page_object.post_name}失败，角色名称已存在')
+                elif not await cls.check_role_key_unique_services(query_db, page_object):
+                    raise ServiceException(message=f'修改角色{page_object.post_name}失败，角色权限已存在')
             try:
                 await RoleDao.edit_role_dao(query_db, edit_role)
                 if page_object.type != 'status':
@@ -114,14 +164,12 @@ class RoleService:
                         for menu in page_object.menu_ids:
                             await RoleDao.add_role_menu_dao(query_db, RoleMenuModel(roleId=page_object.role_id, menuId=menu))
                 await query_db.commit()
-                result = dict(is_success=True, message='更新成功')
+                return CrudResponseModel(is_success=True, message='更新成功')
             except Exception as e:
                 await query_db.rollback()
                 raise e
         else:
-            result = dict(is_success=False, message='角色不存在')
-
-        return CrudResponseModel(**result)
+            raise ServiceException(message='角色不存在')
 
     @classmethod
     async def role_datascope_services(cls, query_db: AsyncSession, page_object: AddRoleModel):
@@ -131,20 +179,9 @@ class RoleService:
         :param page_object: 角色数据权限对象
         :return: 分配角色数据权限结果
         """
-        edit_role = page_object.model_dump(exclude_unset=True, exclude={'admin'})
-        del edit_role['dept_ids']
-        role_info = await cls.role_detail_services(query_db, edit_role.get('role_id'))
-        if role_info:
-            if role_info.role_name != page_object.role_name:
-                role_name = await RoleDao.get_role_by_info(query_db, RoleModel(roleName=page_object.role_name))
-                if role_name:
-                    result = dict(is_success=False, message='角色名称已存在')
-                    return CrudResponseModel(**result)
-            elif role_info.role_key != page_object.role_key:
-                role_key = await RoleDao.get_role_by_info(query_db, RoleModel(roleKey=page_object.role_key))
-                if role_key:
-                    result = dict(is_success=False, message='权限字符已存在')
-                    return CrudResponseModel(**result)
+        edit_role = page_object.model_dump(exclude_unset=True, exclude={'admin', 'dept_ids'})
+        role_info = await cls.role_detail_services(query_db, page_object.role_id)
+        if role_info.role_id:
             try:
                 await RoleDao.edit_role_dao(query_db, edit_role)
                 await RoleDao.delete_role_dept_dao(query_db, RoleDeptModel(roleId=page_object.role_id))
@@ -152,14 +189,12 @@ class RoleService:
                     for dept in page_object.dept_ids:
                         await RoleDao.add_role_dept_dao(query_db, RoleDeptModel(roleId=page_object.role_id, deptId=dept))
                 await query_db.commit()
-                result = dict(is_success=True, message='分配成功')
+                return CrudResponseModel(is_success=True, message='分配成功')
             except Exception as e:
                 await query_db.rollback()
                 raise e
         else:
-            result = dict(is_success=False, message='角色不存在')
-
-        return CrudResponseModel(**result)
+            raise ServiceException(message='角色不存在')
 
     @classmethod
     async def delete_role_services(cls, query_db: AsyncSession, page_object: DeleteRoleModel):
@@ -173,17 +208,20 @@ class RoleService:
             role_id_list = page_object.role_ids.split(',')
             try:
                 for role_id in role_id_list:
+                    role = await cls.role_detail_services(query_db, int(role_id))
+                    if (await RoleDao.count_user_role_dao(query_db, int(role_id))) > 0:
+                        raise ServiceException(message=f'角色{role.role_name}已分配,不能删除')
                     role_id_dict = dict(roleId=role_id, updateBy=page_object.update_by, updateTime=page_object.update_time)
                     await RoleDao.delete_role_menu_dao(query_db, RoleMenuModel(**role_id_dict))
+                    await RoleDao.delete_role_dept_dao(query_db, RoleDeptModel(**role_id_dict))
                     await RoleDao.delete_role_dao(query_db, RoleModel(**role_id_dict))
                 await query_db.commit()
-                result = dict(is_success=True, message='删除成功')
+                return CrudResponseModel(is_success=True, message='删除成功')
             except Exception as e:
                 await query_db.rollback()
                 raise e
         else:
-            result = dict(is_success=False, message='传入角色id为空')
-        return CrudResponseModel(**result)
+            raise ServiceException(message='传入角色id为空')
 
     @classmethod
     async def role_detail_services(cls, query_db: AsyncSession, role_id: int):
@@ -194,7 +232,10 @@ class RoleService:
         :return: 角色id对应的信息
         """
         role = await RoleDao.get_role_detail_by_id(query_db, role_id=role_id)
-        result = RoleModel(**CamelCaseUtil.transform_result(role))
+        if role:
+            result = RoleModel(**CamelCaseUtil.transform_result(role))
+        else:
+            result = RoleModel(**dict())
 
         return result
 
@@ -232,15 +273,16 @@ class RoleService:
         return binary_data
 
     @classmethod
-    async def get_role_user_allocated_list_services(cls, query_db: AsyncSession, page_object: UserRolePageQueryModel, is_page: bool = False):
+    async def get_role_user_allocated_list_services(cls, query_db: AsyncSession, page_object: UserRolePageQueryModel, data_scope_sql: str, is_page: bool = False):
         """
         根据角色id获取已分配用户列表
         :param query_db: orm对象
         :param page_object: 用户关联角色对象
+        :param data_scope_sql: 数据权限对应的查询sql语句
         :param is_page: 是否开启分页
         :return: 已分配用户列表
         """
-        query_user_list = await UserDao.get_user_role_allocated_list_by_role_id(query_db, page_object, is_page)
+        query_user_list = await UserDao.get_user_role_allocated_list_by_role_id(query_db, page_object, data_scope_sql, is_page)
         allocated_list = PageResponseModel(
             **{
                 **query_user_list.model_dump(by_alias=True),
@@ -251,15 +293,16 @@ class RoleService:
         return allocated_list
 
     @classmethod
-    async def get_role_user_unallocated_list_services(cls, query_db: AsyncSession, page_object: UserRolePageQueryModel, is_page: bool = False):
+    async def get_role_user_unallocated_list_services(cls, query_db: AsyncSession, page_object: UserRolePageQueryModel, data_scope_sql: str, is_page: bool = False):
         """
         根据角色id获取未分配用户列表
         :param query_db: orm对象
         :param page_object: 用户关联角色对象
+        :param data_scope_sql: 数据权限对应的查询sql语句
         :param is_page: 是否开启分页
         :return: 未分配用户列表
         """
-        query_user_list = await UserDao.get_user_role_unallocated_list_by_role_id(query_db, page_object, is_page)
+        query_user_list = await UserDao.get_user_role_unallocated_list_by_role_id(query_db, page_object, data_scope_sql, is_page)
         unallocated_list = PageResponseModel(
             **{
                 **query_user_list.model_dump(by_alias=True),
